Make validator handle unknown values
The config file can now flag fields as allowing "unknown", where the
actual "unknown" values can be one of a set of terms that are also
specified in the config.

The dependency checking has to take into account that fields might
contain "unknown" when working out if a dependency has been fulfilled,
so the code gets as bit tortuous there. It's likely to need re-visiting
when it's tried on real-world data, and we'll probably need more tests
to cover all eventualities.