Create a proper type library
B::M::Types has been modified to use MooseX::Types, which allows Moose
types to be specified globally and used more sensibly by "using" that
class.

Modified B::MM::Types to turn it into a proper library. Modified classes
that use Moose types to use that library and the MooseX package that
provides Moose built-in types in the same way. Tweaked a test file to
use the slightly different syntax for the type usage.