Add test for non-CSV files
Previously, reading an excel document would result in a large number of
useless warnings about Unicode, but wouldn't actually cause reading to
fail immediately. Add a crude test to use unix "file" to see if the file
is ASCII and die if not.

Add test excel document and modify test to check for the appropriate
error message.