Fix bug with handling of "if" dependencies
This commit fixes a bug in the validation of fields involved with "if"
relationships.

When validating an "if", the validator looks at the field that defines
the relationship to see if its value is true or false. If it's true, it
goes on to check that one set of fields is valid, if it's false, it
checks the other set of fields.

The bug here arose because the check for "true" or "false" in the "if"
column wasn't explicit; it was testing for perl's version of true ("if (
$x )") rather than a literal "true" (or "yes"). Consequently, both "yes"
and "no" were acting as "true".

This commit fixes the bug by checking explicitly for "true", "yes" and
"1", the values of "true" that are allowed by the Bool validator plugin,
and adds a couple of rows to the test CSV file to check for the right
behaviour.