Remove functionality to handle multiple configs
The checklist module previously understood config files which had
multiple checklists in them. The feature was mainly added for use in
testing, but we've moved on since then and it doesn't make much sense
now. Also, it causes problems with another, later feature, which fixes
noisy Excel CSV files.

This commit removes the code in B::M::Checklist that handled multiple
configurations and updates the tests and test data accordingly.