Check for data in unused columns
When adding rows to a manifest the module checks if the row has more
columns than are expected by the checklist. If so, the unused columns
are thrown away. This commit adds a check that issues a warning if a
removed column contains data. Added a test to check for the warning

Fixed a data file from a different test, which throws up that warning
while checking something else.