Add module to find data in the tracking DBS
This is the first commit of a module that mimics some functionality in
the old Bio::ENA::DataSubmission package. So far it includes the ability
to return VRTrack::Lane objects, given sample or lane IDs.

The tests cover the basic usage but they currently rely on having a
connection to the tracking DBs. That will be changed to make it work
with a local SQLite DB, for obvious reasons.