Add new modules for finding paths and databases
These modules are the bones of a new implementation of the PathFind
modules. They separate off the functionality of finding and connecting
to pathogen tracking databases, and the finding of associated paths on
disk. The two concrete classes share a common Role that handles shared
functionality like reading config files.