Switch type library to use Type::Tiny
Change from using MooseX::Types to Type::Tiny. The latter professes
compatibility with both Moose and Moo, so if we need to switch to Moo
for performance reasons, this is one less thing to have to convert then.
Plus, the changes are trivial and don't require any changes at all in
the classes that use the library.