Fix a test failure on linux
Trying to remove a directory on linux throws an error if the directory
doesn't already exist. Doesn't seem to be a problem on OSX, but it may
be due to a difference in dependency versions. This commit fixes it
anyway, by checking that the directory exists before trying to delete
it...